requirements.txt
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Fri Jan 30 01:34:19 2026

@author: amy
"""

streamlit
google-generativeai